content-security-policy/script-src/script-src-1_4.html

<!DOCTYPE HTML>
<html>
<head>
    <title>eval() should not run without 'unsafe-eval' script-src directive.</title>
    <script src='/resources/testharness.js'></script>
    <script src='/resources/testharnessreport.js'></script>
</head>
<body>
    <h1>eval() should not run without 'unsafe-eval' script-src directive.</h1>
    <div id='log'></div>

   	<script>

        var evalRan = false;

        <!-- This is wrapped in a try because some browsers throw an EvalError. This is neither mandated nor forbidden by the spec, so catch it and only care whether the eval code actually ran. -->
        try {
          eval('evalRan = true;');
        } catch(e) {
          console.log(e);
        }

        test(function() {assert_false(evalRan, "Unsafe eval ran.")}, "eval() should be blocked without an 'unsafe-eval' directive");
    </script>

    <script async defer src='../support/checkReport.sub.js?reportField=violated-directive&reportValue=script-src%20%27self%27+%27unsafe-inline%27'></script>

</body>
</html>
